Move error tessts together
